Take screenshot in dialogs #13184
